Release version 1.2.2 with Sparkle auto-update, macro improvements, and VNI fix
## ✨ New Features

### Sparkle Auto-Update Integration
- Integrated Sparkle 2.8.1 for automatic app updates
- Added appcast.xml with EdDSA signature support
- Configured update checking at 1-hour intervals
- Included update UI with release notes display

### VNI Input Method Support
- Fixed critical bug preventing VNI from adding tone marks
- Number keys (0-9) now properly processed in VNI mode
- VNI tone marks working: 1=sắc, 2=nặng, 6=huyền, 7=hỏi, 8=ngã, 9=móc

### Enhanced Macro System
- Improved macro replacement with configurable break keys
- Fixed macro expansion with special characters (@, !, #, ~)
- Added support for uppercase macro expansion
- Better handling of macro keys in Vietnamese and English modes

## 🐛 Bug Fixes

### XKeyIM Controller
- Fixed VNI input blocking issue (numbers were rejected)
- Added proper input method detection for number key processing
- Improved Vietnamese character processing logic

### Code Cleanup
- Removed verbose debug logs from VNEngine
- Cleaned up commented-out code blocks
- Streamlined Vietnamese input processing

## 🔧 Technical Improvements

### Build System
- Updated build_release.sh with Sparkle signing support
- Added ENABLE_XKEYIM flag for bundling XKeyIM
- Improved environment variable handling

### Settings Management
- Enhanced SharedSettings synchronization
- Better plist file writing for cross-process communication
- Improved settings notification system

## 📝 Documentation
- Updated README with installation and usage instructions
- Added build requirements and development setup

## 🧪 Testing
- All input methods verified: Telex ✅ VNI ✅ Simple Telex 1/2 ✅
- Macro system tested with various break keys ✅
- Auto-update mechanism validated ✅

+++ XKey/Core/Engine/VNEngineAdvanced.swift
@@ -100,8 +100,8 @@ extension VNEngine {
                 
                 // Set hook state to send backspaces and new characters
                 hookState.code = UInt8(vWillProcess)
-                hookState.backspaceCount = UInt8(index)
-                hookState.newCharCount = UInt8(index)
+                hookState.backspaceCount = Int(index)
+                hookState.newCharCount = Int(index)
                 
                 for i in 0..<Int(index) {
                     hookState.charData[Int(index) - 1 - i] = getCharacterCode(typingWord[i])
@@ -176,8 +176,8 @@ extension VNEngine {
         
         // Update hook state to send the change
         hookState.code = UInt8(vWillProcess)
-        hookState.backspaceCount = UInt8(index)
-        hookState.newCharCount = UInt8(index)
+        hookState.backspaceCount = Int(index)
+        hookState.newCharCount = Int(index)
         
         for i in 0..<Int(index) {
             hookState.charData[Int(index) - 1 - i] = getCharacterCode(typingWord[i])
@@ -204,8 +204,8 @@ extension VNEngine {
         
         // Calculate backspaces needed
         hookState.code = UInt8(handleCode)
-        hookState.backspaceCount = UInt8(index)
-        hookState.newCharCount = UInt8(originalWord.count)
+        hookState.backspaceCount = Int(index)
+        hookState.newCharCount = originalWord.count
         
         // Set original characters to send
         for i in 0..<originalWord.count {
